Provide MPSEs with context about the beginning/end of packet processing
An MPSE may be called several times (with different instances) for each
packet being processed. There are situations where it would be useful
for the MPSE class to know when this processing started and ended, so
that searches could potentially be batched together for a single packet.
Enable this by moving the initialisation / final processing of the
match stash to the start and end of fpEvalPacket, and providing calls
to the MPSE to indicate the beginning + end of processing for each
packet.