Add support for overlapping jobs to RXP MPSE plugin
The Titan IC Systems RXP supports a maximum job size of 16k. We can
request searches in packet buffers that are larger than this, so
handle them by submitting overlapping jobs (with an overlap equal
to the maximum fast pattern size), and storing the details of the
offset of each job so we can return the correct `to' value.