Allow use of ARM64 CNTVCT_EL0 register for timing
snort supports the use of rdtsc to get fast, accurate-enough timing on
x86 platforms. The CNTVCT_EL0 register on ARM64 provides a usable
equivalent to userspace code on that platform. It's not the actual
processor clock rate but can vary in accuracy from 1-50Mhz. Its use
gives a ~10% performance improvement on an A53 based platform.