Fix GCC 8 format truncation warnings in DAQ plugins
The buffer passed to strerror_r() is larger than the format buffer
provided by DPE. GCC 8 complains about the potential truncation, so
limit the size of the buffer. Biggest error string in glibc is ~ 50
characters so this should be plenty, but if not we'll gracefully
fallback to just the error number.