Move parameter which might be empty to last position
In Debian builds $ENV{SNORT_PLUGIN_PATH} is empty, and the logic in
generate_built.sh to handle PLUGIN_PATH being empty is broken. An empty
parameter in shell won't cause PLUGIN_PATH to be empty, it takes the value
that is expected for INPUT_FILE and INPUT_FILE takes the value expected for
OUTPUT_FILE and OUTPUT_FILE ends up empty.

This patch moves the parameter which might be empty to the end of the
command.