Fix invalid writes to read-only DAQ shared memory
Typically normalization options are turned off when the DAQ is not
writable (see no_tinit in normalize.cc) so packet modifications and
resizings are disabled.

The exception is in TcpNormalizer where there are attempted calls to
trim_payload with 0 length when inline mode is on, even when
normalization mode is off. Replacing the calls with packet_dropper
allows handling the dropping of packets in this case.

Signed-off-by: Agathiyan Bragadeesh <agathiyan.bragadeesh2@arm.com>